Move LogMessage Payload building from MessageWorker to pre-buffering
